fix: call highlight controller for automod caught message
