+++ src/Application.cpp
@@ -547,7 +547,8 @@ void Application::initPubSub()
                                 senderDisplayName, senderColor};
                             postToThread([chan, action] {
                                 const auto p =
-                                    makeAutomodMessage(action, chan->getName());
+                                    TwitchMessageBuilder::makeAutomodMessage(
+                                        action, chan->getName());
                                 chan->addMessage(p.first);
                                 chan->addMessage(p.second);
 
@@ -579,7 +580,8 @@ void Application::initPubSub()
                 }
 
                 postToThread([chan, action] {
-                    const auto p = makeAutomodMessage(action, chan->getName());
+                    const auto p = TwitchMessageBuilder::makeAutomodMessage(
+                        action, chan->getName());
                     chan->addMessage(p.first);
                     chan->addMessage(p.second);
                 });
@@ -621,7 +623,8 @@ void Application::initPubSub()
                 }
 
                 postToThread([chan, action] {
-                    const auto p = makeAutomodInfoMessage(action);
+                    const auto p =
+                        TwitchMessageBuilder::makeAutomodInfoMessage(action);
                     chan->addMessage(p);
                 });
             });
