+++ src/messages/SharedMessageBuilder.hpp
@@ -57,6 +57,9 @@ protected:
 
     // parseHighlights only updates the visual state of the message, but leaves the playing of alerts and sounds to the triggerHighlights function
     virtual void parseHighlights();
+    static void triggerHighlights(const QString &channelName, bool playSound,
+                                  const std::optional<QUrl> &customSoundUrl,
+                                  bool windowAlert);
 
     void appendChannelName();
